backend/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
funcsigs==1.0.2
Markdown==3.0.1
more-itertools==5.0.0
pathlib2==2.3.3
pluggy==0.9.0
py==1.8.0
Pygments==2.3.1
pytest==4.3.0
python-dateutil==2.8.0
scandir==1.10.0
six==1.12.0
